requirements.txt
streamlit
python-dotenv
pandas
supabase
plotly
